closes #240 
~closes #278~ already closed in https://github.com/prometheus/tsdb/pull/572

combined and refactored the above PRs.

will post results from the benchmarks shortly.

Signed-off-by: Krasi Georgiev <kgeorgie@redhat.com>