@krasi-georgiev 2018-12-10
```
benchcmp no-expand.txt expand.txt
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkPersistedQueries_Series_1M_RESelector_1_Expansion-8              157304440     147421099     -6.28%
BenchmarkPersistedQueries_Series_1M_RESelector_2_Expansion-8              19163203      17982639      -6.16%
BenchmarkPersistedQueries_Series_1M_RESelector_3_Expansion-8              8640811       7914559       -8.40%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_1_Expansion-8      515304845     471832010     -8.44%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_2_Expansion-8      64759526      69743052      +7.70%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_3_Expansion-8      29525021      28743209      -2.65%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_1_Expansion-8     572205794     580150933     +1.39%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_2_Expansion-8     72159910      82287972      +14.04%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_3_Expansion-8     33388668      26196775      -21.54%

benchmark                                                                 old allocs     new allocs     delta
BenchmarkPersistedQueries_Series_1M_RESelector_1_Expansion-8              1000182        1000223        +0.00%
BenchmarkPersistedQueries_Series_1M_RESelector_2_Expansion-8              110304         110325         +0.02%
BenchmarkPersistedQueries_Series_1M_RESelector_3_Expansion-8              40215          40232          +0.04%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_1_Expansion-8      3100695        3100832        +0.00%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_2_Expansion-8      341929         341997         +0.02%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_3_Expansion-8      124624         124676         +0.04%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_1_Expansion-8     3100696        3100830        +0.00%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_2_Expansion-8     341929         341996         +0.02%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_3_Expansion-8     124624         124676         +0.04%

benchmark                                                                 old bytes     new bytes     delta
BenchmarkPersistedQueries_Series_1M_RESelector_1_Expansion-8              55219483      59879260      +8.44%
BenchmarkPersistedQueries_Series_1M_RESelector_2_Expansion-8              6081791       6468332       +6.36%
BenchmarkPersistedQueries_Series_1M_RESelector_3_Expansion-8              2208200       2324264       +5.26%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_1_Expansion-8      144936442     158922606     +9.65%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_2_Expansion-8      15965682      17126937      +7.27%
BenchmarkPersistedQueries_3Blocks_Series_1M_RESelector_3_Expansion-8      5801820       6150190       +6.00%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_1_Expansion-8     144937071     158921465     +9.65%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_2_Expansion-8     15965733      17126544      +7.27%
BenchmarkPersistedQueries_10Blocks_Series_1M_RESelector_3_Expansion-8     5801675       6150330       +6.01%
```