@krasi-georgiev 2019-03-28
ping @codesome 	@gouthamve 	@brian-brazil 	
ready for review (thanks to @VladMasarik for fixing a small bug)