@brian-brazil 2019-03-28
Hmm, I've some benchmarks in #572 also. We should de-duplicate a bit I think.