@krasi-georgiev 2019-06-03
some extensive query benchmarks

cc @gouthamve @codesome