@krasi-georgiev 2019-06-06
@brian-brazil couldn't have a quick look and let me know if you still think we should merge this with `BenchmarkHeadPostingForMatchers`