@brian-brazil 2019-06-06
As I said on IRC there, it'd be best to run the postings benchmarks both ways.