@krasi-georgiev 2019-06-06
What do you mean by `both ways`?