@krasi-georgiev 2019-06-06
These benchmarks also test head and blocks completely separately.
https://github.com/prometheus/tsdb/pull/467/files#diff-a1431a302431f51c7f9e7c0b3cb58ed0R2197

I also don't mind to keep these separately.